remaining js_* controls
