+++ superset-frontend/src/explore/controlPanels/DeckArc.js
@@ -20,8 +20,16 @@ import { t } from '@superset-ui/translation';
 import timeGrainSqlaAnimationOverrides from './timeGrainSqlaAnimationOverrides';
 import { nonEmpty, integer } from '../validators';
 import { columnChoices, PRIMARY_COLOR } from '../controls';
-import { filterNulls, autozoom, dimension, jsColumns } from './Shared_DeckGL';
 import { formatSelectOptions } from '../../modules/utils';
+import {
+  filterNulls,
+  autozoom,
+  dimension,
+  jsColumns,
+  jsDataMutator,
+  jsTooltip,
+  jsOnclickHref,
+} from './Shared_DeckGL';
 
 export default {
   requiresTime: true,
@@ -115,9 +123,9 @@ export default {
       label: t('Advanced'),
       controlSetRows: [
         [jsColumns],
-        ['js_data_mutator'],
-        ['js_tooltip'],
-        ['js_onclick_href'],
+        [jsDataMutator],
+        [jsTooltip],
+        [jsOnclickHref],
       ],
     },
   ],
