+++ superset-frontend/src/explore/controlPanels/DeckHex.js
@@ -21,7 +21,14 @@ import {
   formatSelectOptions,
   formatSelectOptionsForRange,
 } from '../../modules/utils';
-import { filterNulls, autozoom, jsColumns } from './Shared_DeckGL';
+import {
+  filterNulls,
+  autozoom,
+  jsColumns,
+  jsDataMutator,
+  jsTooltip,
+  jsOnclickHref,
+} from './Shared_DeckGL';
 
 export default {
   requiresTime: true,
@@ -77,9 +84,9 @@ export default {
       label: t('Advanced'),
       controlSetRows: [
         [jsColumns],
-        ['js_data_mutator'],
-        ['js_tooltip'],
-        ['js_onclick_href'],
+        [jsDataMutator],
+        [jsTooltip],
+        [jsOnclickHref],
       ],
     },
   ],
