+++ superset-frontend/src/explore/controlPanels/DeckPath.js
@@ -17,7 +17,14 @@
  * under the License.
  */
 import { t } from '@superset-ui/translation';
-import { filterNulls, autozoom, jsColumns } from './Shared_DeckGL';
+import {
+  filterNulls,
+  autozoom,
+  jsColumns,
+  jsDataMutator,
+  jsTooltip,
+  jsOnclickHref,
+} from './Shared_DeckGL';
 
 export default {
   requiresTime: true,
@@ -44,9 +51,9 @@ export default {
       label: t('Advanced'),
       controlSetRows: [
         [jsColumns],
-        ['js_data_mutator'],
-        ['js_tooltip'],
-        ['js_onclick_href'],
+        [jsDataMutator],
+        [jsTooltip],
+        [jsOnclickHref],
       ],
     },
   ],
